# docs/swagger.yaml
basePath: /
definitions:
  api.ConvertResponse:
    properties:
      amount:
        type: string
      converted_amount:
        type: string
      date:
        type: string
      exchange:
        type: string
      first_curr:
        type: string
      second_curr:
        type: string
      source:
        type: string
    type: object
  domain.CurrModel:
    properties:
      code:
        description: Код валюты
        type: string
      date:
        description: Дата получения валюты
        type: string
      name:
        description: Полное название на языке из источника
        type: string
      ratio_buy:
        description: Курс покупки
        type: string
      ratio_sell:
        description: Курс продажи
        type: string
    type: object
  handler.Response:
    properties:
      code:
        description: Код ответа
        type: integer
      data:
        description: Данные
        items: {}
        type: array
      message:
        description: Сообщение для пользователя
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Convertation_service for sources RU,TH
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.htm
  title: Swagger Convertation_service API
  version: "1.0"
paths:
  /convert:
    get:
      description: Конвертация валют в зависимости от источника, требуется предоставление
        двух кодов валют, суммы конвертации, и курса обмена.
      operationId: Convert
      parameters:
      - description: source
        in: path
        name: source
        required: true
        type: string
      - description: first
        in: path
        name: first
        required: true
        type: string
      - description: second
        in: path
        name: second
        required: true
        type: string
      - description: amount
        in: path
        name: amount
        required: true
        type: string
      - description: exchange
        in: path
        name: exchange
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.ConvertResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Конвертация валют
      tags:
      - handlerConvert
  /getAll:
    get:
      consumes:
      - application/json
      description: Получить все валюты из источника. Если источник не указан, берутся
        данные из источника по умолчанию (ЦБ РФ)
      operationId: getAll
      parameters:
      - description: source
        in: path
        name: source
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.CurrModel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Получить все валюты
      tags:
      - GetAll
produces:
- application/json
schemes:
- http
swagger: "2.0"